Fixed the demo. Completely removed the playlist support (since it is now a different library) Updated the README and version
