+++ demo/build.gradle
@@ -4,6 +4,9 @@ dependencies {
     compile project(':library')
     compile 'com.android.support:appcompat-v7:23.1.1'
     compile 'com.squareup.picasso:picasso:2.5.2'
+
+    // Playlist support
+    compile 'com.devbrackets.android:playlistcore:1.0.0-SNAPSHOT'
 }
 
 android {
