+++ demo/src/main/res/layout/video_player_activity.xml
@@ -3,7 +3,7 @@
                 android:layout_width="match_parent"
                 android:layout_height="match_parent">
 
-    <com.devbrackets.android.exomedia.playlist.VideoApi
+    <com.devbrackets.android.exomediademo.playlist.VideoApi
         android:id="@+id/video_play_activity_video_view"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
