+++ library/src/main/java/com/devbrackets/android/exomedia/EMAudioPlayer.java
@@ -31,7 +31,6 @@ import com.devbrackets.android.exomedia.core.listener.ExoPlayerListener;
 import com.devbrackets.android.exomedia.type.MediaSourceType;
 import com.devbrackets.android.exomedia.util.EMDeviceUtil;
 import com.devbrackets.android.exomedia.util.MediaUtil;
-import com.devbrackets.android.playlistcore.event.EMMediaProgressEvent;
 import com.google.android.exoplayer.audio.AudioCapabilities;
 import com.google.android.exoplayer.audio.AudioCapabilitiesReceiver;
 
@@ -59,8 +58,6 @@ public class EMAudioPlayer implements AudioCapabilitiesReceiver.Listener {
 
     private AudioCapabilities audioCapabilities;
     private AudioCapabilitiesReceiver audioCapabilitiesReceiver;
-
-    private EMMediaProgressEvent currentMediaProgressEvent = new EMMediaProgressEvent(0, 0, 0);
 
     public EMAudioPlayer(Context context) {
         this.context = context;
